Sources/Outcompute.Toolkit/Expressions/IsFalseWireExpression.cs
using Outcompute.Toolkit.Expressions.Visitors;

namespace Outcompute.Toolkit.Expressions;

/// <summary>
/// A <see cref="WireExpression"/> that returns whether the expression evaluates to false.
/// </summary>
public sealed record class IsFalseWireExpression : WireExpression
{
    internal IsFalseWireExpression(WireExpression expression)
    {
        Guard.IsNotNull(expression);

        Expression = expression;
    }

    /// <summary>
    /// The expression to convert.
    /// </summary>
    public WireExpression Expression { get; }

    /// <summary>
    /// Makes the specified visitor visit the current expression using the correct overload.
    /// </summary>
    protected internal override WireExpression Accept(WireExpressionVisitor visitor) => visitor.VisitIsFalse(this);
}

public partial record class WireExpression
{
    /// <summary>
    /// Creates a <see cref="WireExpression"/> that returns whether the expression evaluates to false.
    /// </summary>
    public static IsFalseWireExpression IsFalse(WireExpression expression) => new(expression);
}